requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-appconf==1.0.5
django-bootstrap-v5==1.0.7
django-imagekit==4.1.0
django-seed==0.3.1
Faker==9.8.2
idna==3.3
oauthlib==3.1.1
pilkit==2.0
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
urllib3==1.26.7
